+++ TP_API_Wiky_Auth/Services/CommentService.cs
@@ -1,7 +1,8 @@
-﻿using DTOs;
+﻿using DTOs.CommentDTOs;
 using IRepositories;
 using IServices;
 using Models;
+using Models.Exceptions;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -18,19 +19,54 @@ namespace Services
             _commentRepository = commentRepository;
         }
 
-        public async Task<Comment> CreateCommentAsync(Comment comment)
+        public async Task<Comment> CreateCommentAsync(CommentCreateDTO commentDTO, string userId)
         {
-           return await _commentRepository.CreateCommentAsync(comment);
+
+            Comment comment = new Comment
+            {
+                Content = commentDTO.Content,
+                UserId = userId,
+                CreationDate = DateTime.Now,
+                ArticleId = commentDTO.ArticleId,
+            };
+
+            return await _commentRepository.CreateCommentAsync(comment);
         }
 
-        public async Task<bool> DeleteCommentAsync(int commentId)
+        public async Task<Comment> UpdateCommentAsync(CommentUpdateDTO commentUpdateDTO, string userId, bool isAdmin)
         {
-            return await _commentRepository.DeleteCommentAsync(commentId);
+            if (commentUpdateDTO.Content.Length <= 100)
+            {
+                var verifComment = await _commentRepository.GetCommentByUserIdAsync(userId);
+
+                if (verifComment.UserId == userId || isAdmin)
+                {
+
+                    Comment comment = new Comment
+                    {
+                        Id = commentUpdateDTO.Id,
+                        Content = commentUpdateDTO.Content,
+                    };
+                    return await _commentRepository.UpdateCommentAsync(comment);
+                }
+                else throw new MyExceptions();
+            }
+            else throw new Exception("Commentaire trop long, taille max 100 charactères !");
         }
 
-        public async Task<List<Comment>> GetCommentByArticleIdAsync(int articleId)
+        public async Task<bool> DeleteCommentAsync(int commentId, string userId, bool isAdmin)
         {
-            return await _commentRepository.GetCommentByArticleIdAsync(articleId);
+            var verifComment = await _commentRepository.GetCommentByUserIdAsync(userId);
+            if(verifComment.UserId == userId || isAdmin) 
+            {
+                return await _commentRepository.DeleteCommentAsync(commentId);
+            }
+            else throw new MyExceptions();
+        }
+
+        public async Task<List<CommentCreateDTO>> GetCommentsByArticleIdAsync(int articleId)
+        {
+            return await _commentRepository.GetCommentsByArticleIdAsync(articleId);
         }
 
         public async Task<Comment> GetCommentByIdAsync(int commentId)
@@ -38,23 +74,14 @@ namespace Services
             return await _commentRepository.GetCommentByIdAsync(commentId);
         }
 
-        public async Task<List<Comment>> GetCommentByUserIdAsync(string userId)
+        public async Task<Comment> GetCommentByUserIdAsync(string userId)
         {
             return await _commentRepository.GetCommentByUserIdAsync(userId);
         }
 
-        public async Task<Comment> UpdateCommentAsync(CommentUpdateDTO commentUpdateDTO)
+        public async Task<List<Comment>> GetCommentsByUserIdAsync(string userId)
         {
-            if (commentUpdateDTO.Content.Length <= 100)
-            {
-                Comment comment = new Comment
-                {
-                    Id = commentUpdateDTO.Id,
-                    Content = commentUpdateDTO.Content,
-                };
-                return await _commentRepository.UpdateCommentAsync(comment);
-            }
-            else throw new Exception("Commentaire trop long, taille max 100 charactères !");
+            return await _commentRepository.GetCommentsByUserIdAsync(userId);
         }
     }
 }
